[WTEL-5395] : add case_comment graph model + util refactoring

+++ internal/app/catalog.go
@@ -171,8 +171,8 @@ func (s *CatalogService) ListCatalogs(ctx context.Context, req *cases.ListCatalo
 		Context: ctx,
 		Sort:    req.Sort,
 		Fields:  fields,
-		Page:    int(page),
-		Size:    int(req.Size),
+		Page:    int64(page),
+		Size:    int64(req.Size),
 		Time:    t,
 		Filter:  make(map[string]interface{}),
 	}
