+++ internal/app/sla_condition.go
@@ -169,8 +169,8 @@ func (s *SLAConditionService) ListSLAConditions(ctx context.Context, req *cases.
 		Fields:  fields,
 		Context: ctx,
 		Sort:    req.Sort,
-		Page:    int(page),
-		Size:    int(req.Size),
+		Page:    int64(page),
+		Size:    int64(req.Size),
 		Time:    t,
 		Filter:  make(map[string]interface{}),
 	}
